+++ classes/Api/ApiRouter.php
@@ -408,6 +408,7 @@ $r->addRoute('GET', '/gpm/themes/{slug}/field/{type}', [GpmController::class, 'c
         $r->addRoute('GET', '/ping', [SystemController::class, 'ping']);
         $r->addRoute('GET', '/system/environments', [SystemController::class, 'environments']);
         $r->addRoute('POST', '/system/environments', [SystemController::class, 'createEnvironment']);
+        $r->addRoute('DELETE', '/system/environments/{name}', [SystemController::class, 'deleteEnvironment']);
         $r->addRoute('GET', '/system/info', [SystemController::class, 'info']);
         $r->addRoute('DELETE', '/cache', [SystemController::class, 'clearCache']);
         $r->addRoute('GET', '/system/logs/files', [SystemController::class, 'logFiles']);